requirements.txt
Django==1.11.21
pytz==2017.2
psycopg2==2.8.4
python-dotenv==0.9.1